+++ kernel/mm/alloc.c
@@ -7,8 +7,8 @@
 
 /* each allocation has the following structure:
  * ptr: data capacity (allocsizes size)
- * ptr + sizeof(size_t): current size
- * ptr + sizeof(size_t) * 2: data
+ * ptr + sizeof(uint64_t): current size
+ * ptr + sizeof(uint64_t) * 2: data
  * ptr + datasize: poison value
  */
 
@@ -16,10 +16,10 @@
 #define POISON_VALUE            0xdeadbeefbadc0ffel
 #define CACHE_COUNT             12
 
-#define CAPACITY_SIZE(cache)    (cache->size - sizeof(size_t) * 2 \
-                                - USE_POISON * sizeof(size_t))
+#define CAPACITY_SIZE(cache)    (cache->size - sizeof(uint64_t) * 2 \
+                                - USE_POISON * sizeof(uint64_t))
 
-static size_t allocsizes[CACHE_COUNT] =
+static uint64_t allocsizes[CACHE_COUNT] =
 {
     32, 64, 128, 256, 512, 1024, 2048, 4096, 8192, 16384, 32768,
     ALLOC_MAX_SIZE,
@@ -29,24 +29,24 @@ static scache_t *caches[CACHE_COUNT] = {0};
 
 static void initarea(scache_t *cache, void *obj)
 {
-    size_t *ptr = obj;
+    uint64_t *ptr = obj;
     *ptr = CAPACITY_SIZE(cache);
     memset(ptr + 2, 0, *ptr);
 #if USE_POISON == 1
-    *((size_t *)((uintptr_t)obj + cache->size - sizeof(size_t))) = POISON_VALUE;
+    *((uint64_t *)((uintptr_t)obj + cache->size - sizeof(uint64_t))) = POISON_VALUE;
 #endif
 }
 
 static void dtor(scache_t *cache, void *obj)
 {
 #if USE_POISON == 1
-    ASSERT(*(size_t *)((uintptr_t)obj + cache->size - sizeof(size_t))
+    ASSERT(*(uint64_t *)((uintptr_t)obj + cache->size - sizeof(uint64_t))
            == POISON_VALUE);
 #endif
     initarea(cache, obj);
 }
 
-static scache_t *getcachefromsize(size_t size)
+static scache_t *getcachefromsize(uint64_t size)
 {
     scache_t *cache = NULL;
     for (int i = 0; i < CACHE_COUNT; ++i) {
@@ -61,10 +61,10 @@ static scache_t *getcachefromsize(size_t size)
     return cache;
 }
 
-void *alloc(size_t size)
+void *alloc(uint64_t size)
 {
     scache_t *cache = getcachefromsize(size);
-    size_t *ret = slab_allocate(cache);
+    uint64_t *ret = slab_allocate(cache);
     if (ret == NULL)
         return NULL;
     ASSERT(*ret == CAPACITY_SIZE(cache));
@@ -74,18 +74,18 @@ void *alloc(size_t size)
 
 void free(void *ptr)
 {
-    size_t *start = (void*)ptr;
+    uint64_t *start = (void*)ptr;
     start -= 2;
-    size_t size = *start;
+    uint64_t size = *start;
     scache_t *cache = getcachefromsize(size);
     slab_free(cache, start);
 }
 
-void *realloc(void *ptr, size_t size)
+void *realloc(void *ptr, uint64_t size)
 {
-    size_t *start = (void*)ptr;
+    uint64_t *start = (void*)ptr;
     start -= 2;
-    size_t currentsize = *(start + 1);
+    uint64_t currentsize = *(start + 1);
     if (size <= currentsize) {
         *(start + 1) = size;
         return (void*)ptr;
@@ -97,14 +97,14 @@ void *realloc(void *ptr, size_t size)
 
     /* same allocation */
     if (oldcache == newcache) {
-        size_t diff = size - currentsize;
+        uint64_t diff = size - currentsize;
         memset((void *)((uintptr_t)ptr + currentsize), 0, diff);
         *(start + 1) = size;
         return (void*)ptr;
     }
 
     /* different allocation */
-    size_t *new = slab_allocate(newcache);
+    uint64_t *new = slab_allocate(newcache);
     if (new == NULL)
         return NULL;
     *(new + 1) = size;
@@ -116,8 +116,8 @@ void *realloc(void *ptr, size_t size)
 void alloc_init()
 {
     for (int i = 0; i < CACHE_COUNT; ++i) {
-        caches[i] = slab_newcache(allocsizes[i] + sizeof(size_t) * 2
-                                  + sizeof(size_t) * USE_POISON,
+        caches[i] = slab_newcache(allocsizes[i] + sizeof(uint64_t) * 2
+                                  + sizeof(uint64_t) * USE_POISON,
                                   0, initarea, dtor);
         ASSERT(caches[i]);
     }
